Move evaluate javascript on page to evaluation.py, improve docs on wait_for_function
