+++ Browser/keywords/evaluation.py
@@ -1,6 +1,6 @@
 from enum import Enum, auto
 import json
-from typing import Optional
+from typing import Any, Optional
 
 from robotlibcore import keyword  # type: ignore
 
@@ -82,3 +82,16 @@ class Evaluation(LibraryComponent):
                 pass
         logger.info(response)
         return response
+
+    @keyword(
+        name="Execute JavaScript On Page", tags=["Setter", "PageContent", "WebAppState"]
+    )
+    def execute_javascript_on_page(self, script: str) -> Any:
+        """Executes given javascript on the page.
+        """
+        with self.playwright.grpc_channel() as stub:
+            response = stub.ExecuteJavascriptOnPage(
+                Request().JavascriptCode(script=script)
+            )
+            logger.info(response.log)
+            return json.loads(response.result)
