+++ atest/test/03_Waiting/wait_for_function.robot
@@ -5,7 +5,17 @@ Test Setup        New Page    ${LOGIN_URL}
 
 *** Test Cases ***
 Succesful Wait For Function
-    Wait For Function    ()=> true    timeout=500ms
+    Wait For Function    true    timeout=500ms
 
 Failed Wait For Function
-    Run Keyword and Expect Error    Timeout 100ms exceeded during page.waitForFunction*    Wait For Function    ()=>false    timeout=100ms
+    Run Keyword and Expect Error    Timeout 100ms exceeded during page.waitForFunction*    Wait For Function    false    timeout=100ms
+
+Failed Wait For Function Promise
+    ${promise}    Promise To    Wait For Function    document.activeElement === document.querySelector('#username_field')    timeout=100ms
+    Run Keyword and Expect Error    Timeout 100ms exceeded during page.waitForFunction*    Wait For    ${promise}
+    Fail
+
+Wait For Element To Have Focus
+    ${promise}    Promise To    Wait For Function    document.activeElement === document.querySelector('#username_field')    timeout=2s
+    Click    \#username_field
+    Wait For    ${promise}
